+++ pkg/notifier/tofile/client.go
@@ -1,4 +1,4 @@
-package local_file
+package tofile
 
 import (
 	"context"
