+++ pkg/notifier/tofile/comment.go
@@ -1,4 +1,4 @@
-package local_file
+package tofile
 
 import (
 	"context"
@@ -10,27 +10,25 @@ import (
 type LocalFileService service
 
 // Post posts comment
-func (f *LocalFileService) Post(ctx context.Context, body string, OutputFile string) error {
+func (f *LocalFileService) Post(ctx context.Context, body string, outputFile string) error {
 	logE := logrus.WithFields(logrus.Fields{
 		"program": "tfcmt",
 	})
 
-	
-	file, err := os.Create(OutputFile)
-	
+	file, err := os.Create(outputFile)
+
 	if err != nil {
 		return err
 	}
 
-
 	defer file.Close()
-	
+
 	_, err2 := file.WriteString(body)
-	
+
 	if err2 != nil {
 		return err2
 	}
 	logE.Debug("Output to file success")
-	
+
 	return err
 }
